Document some tests to make them easier to understand.
